+++ src/norm/private/sqlite/rowutils.nim
@@ -31,10 +31,13 @@ proc fromRowPos[T: Model](obj: var T, row: Row, pos: var Natural, skip = false) 
           subMod.fromRowPos(row, pos, skip = true)  ## Then we skip all the ``row`` values that should've gone into this submodel.
 
         else:                                       ## If ``row[pos]`` is not a ``NULL``,
-          inc pos
+          inc pos                                   ##
           subMod.fromRowPos(row, pos)               ## we actually populate the submodel.
 
-    elif not skip:                                  ## If we're dealing with an "orginary" field,
+      else:                                         ## If the field is a ``none Model``,
+        inc pos                                     ## don't bother trying to populate it at all.
+
+    elif not skip:                                  ## If we're dealing with an "ordinary" field,
       val = row[pos].to(typeof(val))                ## just convert it.
       inc pos