+++ tests/sqlite/trows.nim
@@ -26,16 +26,29 @@ suite "Row CRUD":
     removeFile dbFile
 
   test "Insert row":
-    var toy = newToy(123.45)
+    var toy = newtoy(123.45)
 
-    dbConn.insert(toy)
+    dbconn.insert(toy)
 
     check toy.id > 0
 
-    let rows = dbConn.getAllRows(sql"SELECT price, id FROM Toy")
+    let rows = dbconn.getallrows(sql"select price, id from toy")
 
     check rows.len == 1
     check rows[0] == @[?123.45, ?toy.id]
+
+  test "Insert row twice":
+    var toy = newtoy(123.45)
+
+    dbconn.insert(toy)
+    dbconn.insert(toy)
+
+    check toy.id > 1
+
+    let rows = dbconn.getallrows(sql"select price, id from toy")
+
+    check rows.len == 2
+    check rows[^1] == @[?123.45, ?toy.id]
 
   test "Insert rows":
     var person = newPerson("Alice", newPet("cat", newToy(123.45)))
@@ -72,6 +85,17 @@ suite "Row CRUD":
       select(outToy, "price = ?", inpToy.price)
 
     check outToy === inpToy
+
+  test "Get row, record not found":
+    var outToy = newToy()
+
+    expect KeyError:
+      with dbConn:
+        select(outToy, "Toy.id = ?", 1)
+
+    expect NotFoundError:
+      with dbConn:
+        select(outToy, "Toy.id = ?", 1)
 
   test "Get row, no intermediate objects":
     let